+++ main/VisualItem.h
@@ -6,11 +6,11 @@ class VisualItem
 {
 
 public:
-    VisualItem(wxString name, int startTime, int endTime, bool isStatus = false, bool isDoubleQueue = false);
+    VisualItem(wxString name, double startTime, double endTime, bool isStatus = false, bool isDoubleQueue = false);
 
     wxString name;
-    int startTime;
-    int endTime;
+    double startTime;
+    double endTime;
     bool isStatus;
     bool isDoubleQueue;
 };
